Move dimension entity to model
