+++ pom.xml
@@ -40,7 +40,7 @@
         <dependency>
             <groupId>com.github.ONSdigital</groupId>
             <artifactId>dp-dd-backend-model</artifactId>
-            <version>develop-SNAPSHOT</version>
+            <version>feature~add_dimension_table-SNAPSHOT</version>
         </dependency>
         <!--<dependency>-->
             <!--<groupId>uk.co.onsdigital.discovery</groupId>-->
