+++ src/main/resources/META-INF/persistence.xml
@@ -7,6 +7,7 @@
         <class>uk.co.onsdigital.discovery.model.DimensionalDataPoint</class>
         <class>uk.co.onsdigital.discovery.model.DimensionalDataPointPK</class>
         <class>uk.co.onsdigital.discovery.model.DimensionalDataSet</class>
+        <class>uk.co.onsdigital.discovery.model.Dimension</class>
         <class>uk.co.onsdigital.discovery.model.DimensionValue</class>
         <class>uk.co.onsdigital.discovery.model.GeographicArea</class>
         <class>uk.co.onsdigital.discovery.model.GeographicAreaHierarchy</class>
@@ -28,8 +29,6 @@
         <class>uk.co.onsdigital.discovery.model.UUIDConverter</class>
         <class>uk.co.onsdigital.discovery.model.ValueDomain</class>
         <class>uk.co.onsdigital.discovery.model.Variable</class>
-
-        <class>uk.co.onsdigital.discovery.metadata.api.model.Dimension</class>
         <exclude-unlisted-classes>false</exclude-unlisted-classes>
         <properties>
             <property name="javax.persistence.jdbc.url" value="jdbc:postgresql://localhost:5432/data_discovery" />
@@ -40,11 +39,6 @@
             <!-- Disable the EclipseLink shared entity cache as we expect another service (the loader) to be changing
                  the database directly. -->
             <property name="eclipselink.cache.shared.default" value="false"/>
-
-            <!-- Run scripts on startup to add additional views that the API needs -->
-            <property name="javax.persistence.schema-generation.database.action" value="create"/>
-            <property name="javax.persistence.schema-generation.create-source" value="script"/>
-            <property name="javax.persistence.schema-generation.create-script-source" value="META-INF/views.sql" />
 
             <!-- Show generated SQL -->
             <property name="eclipselink.logging.level.sql" value="FINE"/>
